break up lambda module, replace bunyan with minlog
