+++ package.json
@@ -11,14 +11,12 @@
   ],
   "dependencies": {
     "aws-sdk": "^2.205.0",
-    "bunyan": "^1.8.12",
-    "bunyan-format": "^0.2.1",
-    "bunyan-slack": "^0.0.10",
     "cors": "^2.8.4",
     "express": "^4.16.2",
     "express-bearer-token": "^2.1.1",
     "http-lambda": "git://github.com/tobiipro/http-lambda.git#semver:~0.2.0",
     "lodash-firecloud": "git://github.com/tobiipro/lodash-firecloud.git#semver:~0.2.1",
+    "minlog": "git://github.com/tobiipro/minlog.git#semver:~0.2.4",
     "response-time": "^2.3.2"
   },
   "devDependencies": {
