+++ src/express/res-error.js
@@ -35,7 +35,7 @@ export let bootstrap = function(fn, res) {
         return res.sendError(err);
       }
 
-      if (res.ctx.log.level() <= res.ctx.log.resolveLevel('TRACE')) {
+      if (res.ctx.log.level() >= res.ctx.log.levelToLevelCode('TRACE')) {
         ctx.log.info('Responding with trace...');
         let internalErr = new ResponseError(500, {
           renderer: pkg.name,
